requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pdf2image>=1.16.0
paddlepaddle
paddleocr
easyocr>=1.7.0
python-dateutil>=2.8.0

streamlit-option-menu>=0.3.6
jsonschema>=4.19.0

